update add feedback table, add sentiment to product and add statistic Sentiment by shop

+++ gt-backend/src/main/java/gt/electronic/ecommerce/controller/BrandController.java
@@ -54,7 +54,7 @@ public class BrandController {
 
   @GetMapping("/shopId/{shopId}")
   public ResponseObject<List<BrandResponseDTO>> getAllBrandsByShop(
-      @PathVariable(name = "shopId") Integer shopId) {
+      @PathVariable(name = "shopId") Long shopId) {
     return new ResponseObject<>(
         HttpStatus.OK, "", this.brandService.getAllBrandsByShop(shopId));
   }
