+++ gt-backend/src/main/java/gt/electronic/ecommerce/services/CartItemService.java
@@ -15,17 +15,17 @@ import java.util.List;
  * @created 11/09/2022 - 10:00 AM
  */
 public interface CartItemService {
-  List<OrderDetailResponseDTO> getAllCartItemByUser(String loginKey, Integer userId, Pageable pageable);
+  List<OrderDetailResponseDTO> getAllCartItemByUser(String loginKey, Long userId, Pageable pageable);
 
   OrderDetailResponseDTO getCartItemById(Long id);
 
-  OrderDetailResponseDTO getCartItemByUserAndProduct(Integer userId, Long productId);
+  OrderDetailResponseDTO getCartItemByUserAndProduct(Long userId, Long productId);
 
   OrderDetailResponseDTO createCartItem(String loginKey, CartItemCreationDTO creationDTO);
 
   OrderDetailResponseDTO updateCartItem(String loginKey, Long id, CartItemCreationDTO creationDTO);
 
-  OrderDetailResponseDTO deleteCartItemById(String loginKey, Integer userId, Long id);
+  OrderDetailResponseDTO deleteCartItemById(String loginKey, Long userId, Long id);
 
   List<OrderDetailResponseDTO> updateCart(String loginKey, List<CartDetailCreationDTO> creationDTOList);
 }
